# Implement ToNapiValue trait for RowWrapper