--- a/lib/types/results-wrapper.js
+++ b/lib/types/results-wrapper.js
@@ -100,7 +100,7 @@ function getRowsFromResultsWrapper(result) {
     let rows = [];
 
     for (let i = 0; i < rustRows.length; i++) {
-        let cols = rustRows[i].getColumns();
+        let cols = rustRows[i];
         let collectedRow = new Row(colNames);
         for (let j = 0; j < cols.length; j++) {
             // By default driver returns row as a map: