--- a/src/result.rs
+++ b/src/result.rs
@@ -8,7 +8,7 @@ use crate::{
     utils::err_to_napi,
 };
 use napi::{
-    bindgen_prelude::{BigInt, Buffer},
+    bindgen_prelude::{BigInt, Buffer, ToNapiValue},
     Error, Status,
 };
 use scylla::{
@@ -36,7 +36,9 @@ pub struct QueryResultWrapper {
 }
 
 /// Wrapper for a single row of the query result
-#[napi]
+///
+/// Whenever it's returned from NAPI function call
+/// it's automatically converted to the inner vector
 pub struct RowWrapper {
     inner: Vec<Option<CqlValue>>,
 }
@@ -149,17 +151,18 @@ impl QueryResultWrapper {
     }
 }
 
-#[napi]
-impl RowWrapper {
-    #[napi]
-    /// Get the CQL value wrappers for each column in the given row
-    pub fn get_columns(&self) -> napi::Result<Vec<Option<CqlValueWrapper>>> {
-        let s: Vec<Option<CqlValueWrapper>> = self
-            .inner
-            .iter()
-            .map(|f| f.clone().map(|f| CqlValueWrapper { inner: f }))
-            .collect();
-        Ok(s)
+impl ToNapiValue for RowWrapper {
+    unsafe fn to_napi_value(
+        env: napi::sys::napi_env,
+        val: Self,
+    ) -> napi::Result<napi::sys::napi_value> {
+        Vec::to_napi_value(
+            env,
+            val.inner
+                .into_iter()
+                .map(|e| e.map(|f| CqlValueWrapper { inner: f }))
+                .collect(),
+        )
     }
 }